Generators for tests moved to separate trait
