+++ src/test/scala/ch/zzeekk/spark/temporalquery/TestUtils.scala
@@ -197,53 +197,4 @@ trait TestUtils extends Logging {
     (1, "2020-01-15 09:00:00", finisTemporisString, "2019-03-03 01:00:00", "2021-12-01 02:34:56.1", -2.0)
   ).map(makeRowsBiTemporal).toDF("id", "known_from", "known_to", "valid_from", "valid_to", "value")
 
-  /////////////////////////////////////////////////////////////////////
-  ///// generator of data frames with 2 dimensions of type Double /////
-  /////////////////////////////////////////////////////////////////////
-  val testUnitInterval: Gen[Double] = choose[Double](min = 0.1, max = 0.9)
-
-  case class Point(x: Double = 0d, y: Double = 0d) extends Ordered[Point] with Serializable {
-    def compare(that: Point): Int = this.x.compare(that.x) match {
-      case 0 => this.y.compare(that.y)
-      case n => n
-    }
-  }
-
-  case class Rectangle(lowerLeftPoint: Point = Point(), upperRightPoint: Point = Point(x = 1d, y = 1d)) {
-    def isPtInside(pt: Point): Boolean = lowerLeftPoint.x < pt.x && lowerLeftPoint.y < pt.y &&
-      pt.x < upperRightPoint.x && pt.y < upperRightPoint.y
-  }
-
-  val testUnitPoint2: Gen[Point] = Gen
-    .zip[Double, Double](g1 = testUnitInterval, g2 = testUnitInterval).map { case (x, y) => Point(x, y) }
-  val testUnitPoint2s: Gen[List[Point]] = nonEmptyListOf(g = testUnitPoint2)
-
-  def splitRectangle(rect: Rectangle = Rectangle())(splitPoints: Seq[Point]): List[Rectangle] = {
-    val splitPtsInside = splitPoints.filter(rect.isPtInside).sorted
-    if (splitPtsInside.isEmpty) List(rect)
-    else {
-      val splitPt = splitPtsInside.head
-      splitRectangle(Rectangle(rect.lowerLeftPoint, splitPt))(splitPtsInside.tail) ++
-        splitRectangle(Rectangle(Point(rect.lowerLeftPoint.x, splitPt.y), Point(splitPt.x, rect.upperRightPoint.y)))(splitPtsInside.tail) ++
-        splitRectangle(Rectangle(Point(splitPt.x, rect.lowerLeftPoint.y), Point(rect.upperRightPoint.x, splitPt.y)))(splitPtsInside.tail) ++
-        splitRectangle(Rectangle(splitPt, rect.upperRightPoint))(splitPtsInside.tail)
-    }
-  }
-
-  def rectangleValue2RowTuple[T](value: T)(rect: Rectangle): (Double, Double, Double, Double, T) =
-    (rect.lowerLeftPoint.x, rect.upperRightPoint.x, rect.lowerLeftPoint.y, rect.upperRightPoint.y, value)
-
-  /**
-   * generates data frames with two dense dimensions of which the domain is the unit square with
-   * constant string value
-   *
-   * @param value
-   *   the constant string value
-   * @return
-   *   dataFrame with several entries with constant value
-   */
-  def dfConstantUnitSplitted(value: String): Gen[DataFrame] = testUnitPoint2s.map(splitRectangle())
-    .map[List[(Double, Double, Double, Double, String)]](rects => rects.map(rectangleValue2RowTuple(value)))
-    .map(_.toDF("x_from", "x_to", "y_from", "y_to", "value"))
-
 }
